Hotfix OpenIdCredentials

See merge request famedly/famedlysdk!342

+++ lib/matrix_api/model/open_id_credentials.dart
@@ -20,7 +20,7 @@ class OpenIdCredentials {
   String accessToken;
   String tokenType;
   String matrixServerName;
-  int expiresIn;
+  double expiresIn;
 
   OpenIdCredentials.fromJson(Map<String, dynamic> json) {
     accessToken = json['access_token'];
